Applied modular grappling hook stats

+++ src/main/java/de/bax/dysonsphere/capabilities/grapplingHook/IGrapplingHookRope.java
@@ -8,7 +8,7 @@ import net.minecraftforge.common.capabilities.AutoRegisterCapability;
 @AutoRegisterCapability
 public interface IGrapplingHookRope {
     
-    public double getMaxDistance(Level level, Player player);
+    public float getMaxDistance(Level level, Player player);
 
     public float getDeployForceMultiplier(Level level, Player player);
